Refactor: Remove MakeItNaturalProps type and update imports
