+++ src/features/make-it-natural/api/index.ts
@@ -1,3 +1,2 @@
 export { useMakeItNaturalForm } from "./use-make-it-natural-form";
 export { useMakeItNaturalQuery } from "./queries";
-export type { MakeItNaturalProps } from "./props";
